chore(broker): make partition id unique
- with the system topic we can ensure uniqueness and consistency
  of partition ids
- removes the need to identify a partition via topic name and partition
  id
- removes topic name from client requests where it is not required
- removes the default topic

related to #410, #414

+++ broker-core/src/main/java/io/zeebe/broker/clustering/handler/TopicLeader.java
@@ -17,9 +17,9 @@
  */
 package io.zeebe.broker.clustering.handler;
 
-import io.zeebe.msgpack.UnpackedObject;
 import org.agrona.DirectBuffer;
 
+import io.zeebe.msgpack.UnpackedObject;
 import io.zeebe.msgpack.property.IntegerProperty;
 import io.zeebe.msgpack.property.StringProperty;
 
@@ -73,7 +73,7 @@ public class TopicLeader extends UnpackedObject
         return this;
     }
 
-    public int getPartitionId()
+    public long getPartitionId()
     {
         return partitionIdProp.getValue();
     }
